requirements.txt
duckdb
pandas
requests
python-dotenv
ipykernel
importlib